Implement a staging area for motion settings configured for the next trajectory.
Incoming motion settings are configured before a new trajectory is
loaded (e.g. with services). This is a no-op in the base
JointTrajectoryController, but derived classes may implement something
here.

The "queued" settings instance is only for non-realtime use, and is
copied into the trajectory structure just before the trajectory swap.
Note that even if the update fails, any queued settings are cleared out.